simplify numeric types & conversions

+++ src/procgen.rs
@@ -118,7 +118,7 @@ where
 
 /// Generates a random number using the cell and the given seed s/t the number is the same for each cell.
 /// This ensures that a specific cell will yield the same random result as long as `seed` is the same.
-pub fn stable_hash(cell: &Cell, seed: u64) -> usize {
+pub fn stable_hash(cell: &Cell, seed: u64) -> u32 {
     use rand::rngs::StdRng;
     use rand::{Rng, SeedableRng};
     use std::hash::{DefaultHasher, Hash, Hasher};
@@ -129,7 +129,7 @@ pub fn stable_hash(cell: &Cell, seed: u64) -> usize {
     let hash = hasher.finish();
 
     let mut rng = StdRng::seed_from_u64(hash);
-    rng.next_u32() as usize
+    rng.next_u32()
 }
 
 #[cfg(test)]
@@ -147,7 +147,7 @@ mod tests {
     #[test]
     fn stable_hash_different_cells_differ() {
         // Collect hashes for a small grid and assert they're not all identical.
-        let hashes: Vec<usize> = (0..5)
+        let hashes: Vec<u32> = (0..5)
             .flat_map(|x| (0..5).map(move |y| stable_hash(&Cell::new(x, y), 1)))
             .collect();
         let first = hashes[0];
